backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
django-filter==24.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-dotenv==1.0.1
djangorestframework-camel-case==1.4.2
Pillow==10.3.0
certifi==2024.6.2 